change idProduit type

+++ src/main/java/fr/miage/agents/api/message/negociation/NegocierPrix.java
@@ -12,7 +12,7 @@ public class NegocierPrix extends Message {
 
     public UUID session;
 
-    public int idProduit;
+    public long idProduit;
 
     public float prixDemande;
